refactored firebase code and added achievements view for guest users
